@@ -197,88 +206,88 @@
     "timeago.js": "4.0.2",
     "uuid": "9.0.0",
     "winston": "3.8.2",
-    "yargs": "17.6.0"
+    "yargs": "17.7.1"
   },
   "devDependencies": {
-    "@babel/core": "7.19.6",
+    "@babel/core": "7.21.0",
     "@babel/preset-env": "7.20.2",
-    "@commitlint/cli": "17.1.2",
-    "@commitlint/config-conventional": "17.1.0",
+    "@commitlint/cli": "17.4.4",
+    "@commitlint/config-conventional": "17.4.4",
     "@ffmpeg-installer/ffmpeg": "1.1.0",
-    "@types/chai": "4.3.3",
-    "@types/express": "4.17.14",
+    "@types/chai": "4.3.4",
+    "@types/express": "4.17.17",
     "@types/i18next-fs-backend": "1.1.2",
-    "@types/mocha": "10.0.0",
-    "@types/node": "18.11.7",
+    "@types/mocha": "10.0.1",
+    "@types/node": "18.14.0",
     "@types/proxyquire": "1.3.28",
     "@types/semver-sort": "0.0.1",
     "@types/sinon": "10.0.13",
-    "@types/sinon-chai": "3.2.8",
-    "@typescript-eslint/eslint-plugin": "5.42.0",
-    "@typescript-eslint/parser": "5.42.0",
-    "babel-loader": "8.2.5",
+    "@types/sinon-chai": "3.2.9",
+    "@typescript-eslint/eslint-plugin": "5.51.0",
+    "@typescript-eslint/parser": "5.53.0",
+    "babel-loader": "9.1.2",
     "babel-plugin-syntax-dynamic-import": "6.18.0",
-    "browser-sync": "2.27.10",
+    "browser-sync": "2.27.11",
     "browser-sync-webpack-plugin": "2.3.0",
-    "chai": "4.3.6",
+    "chai": "4.3.7",
     "chai-as-promised": "7.1.1",
     "cheerio": "1.0.0-rc.12",
     "copy-webpack-plugin": "9.1.0",
     "coveralls": "3.1.1",
-    "css-loader": "6.7.1",
+    "css-loader": "6.7.3",
     "del-cli": "5.0.0",
-    "eslint": "^8.26.0",
-    "eslint-config-prettier": "8.5.0",
+    "eslint": "^8.34.0",
+    "eslint-config-prettier": "8.6.0",
     "eslint-config-standard": "17.0.0",
-    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-import": "2.27.5",
     "eslint-plugin-mocha": "10.1.0",
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prettier": "4.2.1",
     "eslint-plugin-promise": "6.1.1",
     "eslint-plugin-testcafe": "0.2.1",
     "finepack": "2.10.15",
-    "git-authors-cli": "1.0.44",
-    "glob": "8.0.3",
-    "husky": "8.0.1",
+    "git-authors-cli": "1.0.45",
+    "glob": "8.1.0",
+    "husky": "8.0.3",
     "image-minimizer-webpack-plugin": "2.2.0",
     "imagemin-gifsicle": "7.0.0",
     "imagemin-jpegtran": "7.0.0",
     "imagemin-optipng": "8.0.0",
     "imagemin-svgo": "10.0.1",
-    "lint-staged": "13.0.3",
-    "mini-css-extract-plugin": "2.6.1",
-    "mocha": "10.1.0",
+    "lint-staged": "13.1.2",
+    "mini-css-extract-plugin": "2.7.2",
+    "mocha": "10.2.0",
     "mocha-junit-reporter": "2.2.0",
-    "mock-fs": "5.1.4",
-    "nock": "13.2.9",
+    "mock-fs": "5.2.0",
+    "nock": "13.3.0",
     "nodemon": "2.0.20",
-    "npm-check-updates": "16.3.22",
+    "npm-check-updates": "16.7.9",
     "nyc": "15.1.0",
     "open-cli": "7.1.0",
     "pinst": "3.0.0",
-    "prettier": "2.7.1",
+    "prettier": "2.8.4",
     "proxyquire": "2.1.3",
-    "sass": "1.56.1",
+    "sass": "1.58.3",
     "sass-loader": "^13.2.0",
-    "sinon": "14.0.1",
+    "sinon": "15.0.1",
     "sinon-chai": "3.7.0",
     "standard": "17.0.0",
     "standard-version": "9.5.0",
     "terser-webpack-plugin": "5.3.6",
-    "testcafe": "2.0.1",
-    "testcafe-browser-provider-browserstack": "1.14.0",
-    "testcafe-reporter-xunit": "2.2.1",
-    "timezone-mock": "1.3.4",
+    "testcafe": "2.3.1",
+    "testcafe-browser-provider-browserstack": "1.14.1",
+    "testcafe-reporter-xunit": "2.2.2",
+    "timezone-mock": "1.3.6",
     "ts-node": "10.9.1",
-    "typescript": "4.8.4",
-    "webpack": "5.74.0",
+    "typescript": "4.9.5",
+    "webpack": "5.75.0",
     "webpack-assets-manifest": "5.1.0",
     "webpack-cli": "4.6.0",
     "webpack-merge": "5.7.3"
   },
   "engines": {
-    "node": "16",
-    "npm": "8"
+    "node": "18",
+    "npm": "9"
   },
   "files": [],
   "scripts": {
