@@ -1,7 +1,7 @@
 const localsActions = require('./locals.actions')
 const localsIdentityBar = require('./locals.identity-bar')
 const localsMessageBanner = require('./locals.message-banner')
-const localsMoveDetails = require('./locals.move-details')
+const localsMoveDetails = require('./locals.move-details').default
 const localsPersonSummary = require('./locals.person-summary')
 const localsTabs = require('./locals.tabs')
 const localsUrls = require('./locals.urls')
