@@ -97,9 +97,9 @@ describe('Police Custody Form controllers', function () {
           PerSelfHarm: '',
           PerEscape: '',
           PersonMoveUsedForce: '',
-          PerMedicalAid1: '',
+          PerMedicalAid: '',
           PerMedicalDrugsAlcohol: '',
-          PerMedicalAid2: '',
+          PerMedicalMedication: '',
           PerMedicalMentalHealth: '',
           PerPropertyChange: '',
           PersonMoveDeathInCustody: '',
