@@ -1,11 +1,16 @@
 const i18n = require('../../config/i18n').default
 const filters = require('../../config/nunjucks/filters')
+const mapUpdateLink = require('../helpers/move/map-update-link')
 
-function moveToAdditionalInfoListComponent({
-  move_type: moveType,
-  time_due: timeDue,
-  additional_information: additionalInformation,
-} = {}) {
+function moveToAdditionalInfoListComponent(
+  {
+    move_type: moveType,
+    time_due: timeDue,
+    additional_information: additionalInformation,
+    recall_date: recallDate,
+  } = {},
+  updateUrls = []
+) {
   if (!['hospital', 'prison_recall', 'video_remand'].includes(moveType)) {
     return undefined
   }
@@ -31,8 +36,39 @@ function moveToAdditionalInfoListComponent({
         classes: !additionalInformation ? 'app-secondary-text-colour' : '',
         text: additionalInformation || i18n.t('not_provided'),
       },
+      updateJourneyKey: moveType === 'prison_recall' ? 'move' : null,
     },
-  ].filter(row => row.value.text || row.value.html)
+    {
+      key: {
+        text: i18n.t('recall_date', {
+          context: moveType,
+        }),
+      },
+      value: {
+        classes: !recallDate ? 'app-secondary-text-colour' : '',
+        text: filters.formatDateWithRelativeDay(recallDate),
+      },
+      updateJourneyKey: 'recall_info',
+    },
+  ]
+    .filter(row => row.value.text || row.value.html)
+    .map(row => {
+      if (!updateUrls[row.updateJourneyKey]) {
+        return row
+      }
+
+      return {
+        ...row,
+        actions: {
+          items: [
+            mapUpdateLink(
+              updateUrls[row.updateJourneyKey],
+              row.updateJourneyKey
+            ),
+          ],
+        },
+      }
+    })
 
   return {
     classes: 'govuk-!-font-size-16',
