@@ -3,16 +3,19 @@ const timezoneMock = require('timezone-mock')
 const i18n = require('../../config/i18n').default
 const filters = require('../../config/nunjucks/filters')
 
-const moveToSummaryListComponent = require('./move-to-additional-info-list-component')
+const moveToAdditionalInfoListComponent = require('./move-to-additional-info-list-component')
 
 const mockMove = {
   time_due: '2000-01-01T14:00:00Z',
   additional_information: 'Some more info',
   move_type: 'hospital',
+  recall_date: '2023-01-09',
 }
 
+const formattedDate = 'Monday 9 Jan 2023'
+
 describe('Presenters', function () {
-  describe('#moveToSummaryListComponent()', function () {
+  describe('#moveToAdditionalInfoListComponent()', function () {
     let transformedResponse
 
     beforeEach(function () {
@@ -28,7 +31,7 @@ describe('Presenters', function () {
 
     context('when provided no move', function () {
       beforeEach(function () {
-        transformedResponse = moveToSummaryListComponent()
+        transformedResponse = moveToAdditionalInfoListComponent()
       })
 
       it('should return undefined', function () {
@@ -38,13 +41,13 @@ describe('Presenters', function () {
 
     context('when provided with a mock move object', function () {
       beforeEach(function () {
-        transformedResponse = moveToSummaryListComponent(mockMove)
+        transformedResponse = moveToAdditionalInfoListComponent(mockMove)
       })
 
       describe('response', function () {
         it('should contain correct rows', function () {
           expect(transformedResponse).to.have.property('rows')
-          expect(transformedResponse.rows.length).to.equal(2)
+          expect(transformedResponse.rows.length).to.equal(3)
         })
 
         it('should order items correctly', function () {
@@ -52,6 +55,7 @@ describe('Presenters', function () {
           expect(keys).to.deep.equal([
             'fields::time_due.label',
             'fields::additional_information.display.label',
+            'recall_date',
           ])
         })
 
@@ -61,12 +65,13 @@ describe('Presenters', function () {
           expect(keys).to.deep.equal([
             mockMove.time_due,
             mockMove.additional_information,
+            formattedDate,
           ])
         })
 
         it('should contain correct row structure', function () {
           transformedResponse.rows.forEach(row => {
-            expect(row).to.have.all.keys(['key', 'value'])
+            expect(row).to.include.keys(['key', 'value'])
           })
         })
 
@@ -77,7 +82,7 @@ describe('Presenters', function () {
 
         it('should contain count', function () {
           expect(transformedResponse).to.have.property('count')
-          expect(transformedResponse.count).to.equal(2)
+          expect(transformedResponse.count).to.equal(3)
         })
 
         it('should contain heading', function () {
@@ -116,7 +121,7 @@ describe('Presenters', function () {
 
     context('with missing additional information', function () {
       beforeEach(function () {
-        transformedResponse = moveToSummaryListComponent({
+        transformedResponse = moveToAdditionalInfoListComponent({
           ...mockMove,
           additional_information: null,
         })
@@ -127,14 +132,18 @@ describe('Presenters', function () {
           const keys = transformedResponse.rows.map(
             row => row.value.text || row.value.html
           )
-          expect(keys).to.deep.equal([mockMove.time_due, 'not_provided'])
+          expect(keys).to.deep.equal([
+            mockMove.time_due,
+            'not_provided',
+            formattedDate,
+          ])
         })
       })
     })
 
     context('with missing time due', function () {
       beforeEach(function () {
-        transformedResponse = moveToSummaryListComponent({
+        transformedResponse = moveToAdditionalInfoListComponent({
           ...mockMove,
           time_due: null,
         })
@@ -145,7 +154,109 @@ describe('Presenters', function () {
           const keys = transformedResponse.rows.map(
             row => row.value.text || row.value.html
           )
-          expect(keys).to.deep.equal([mockMove.additional_information])
+          expect(keys).to.deep.equal([
+            mockMove.additional_information,
+            formattedDate,
+          ])
+        })
+      })
+    })
+
+    context('with missing recall date', function () {
+      beforeEach(function () {
+        transformedResponse = moveToAdditionalInfoListComponent({
+          ...mockMove,
+          recall_date: null,
+        })
+      })
+
+      describe('response', function () {
+        it('should not include recall date', function () {
+          const keys = transformedResponse.rows.map(
+            row => row.value.text || row.value.html
+          )
+          expect(keys).to.deep.equal([
+            mockMove.time_due,
+            mockMove.additional_information,
+          ])
+        })
+      })
+    })
+
+    context('with update URLs', function () {
+      const updateUrls = {
+        move: 'http://www.example.com/move-details',
+        recall_info: 'http://www.example.com/recall-info',
+      }
+
+      context('when prison recall move', function () {
+        beforeEach(function () {
+          transformedResponse = moveToAdditionalInfoListComponent(
+            {
+              ...mockMove,
+              time_due: null,
+              move_type: 'prison_recall',
+            },
+            updateUrls
+          )
+        })
+
+        describe('response', function () {
+          it('should include the update link for recall date', function () {
+            const recallDateRow = transformedResponse.rows.find(
+              row => row.updateJourneyKey === 'recall_info'
+            )
+
+            expect(recallDateRow.actions.items).to.eql([
+              {
+                attributes: {
+                  'data-update-link': 'recall_info',
+                },
+                category: 'recall_info',
+                href: 'http://www.example.com/recall-info',
+                html: 'moves::update_link.link_text',
+              },
+            ])
+          })
+
+          it('should include the update link for additional info', function () {
+            const additionalInfoRow = transformedResponse.rows.find(
+              row => row.updateJourneyKey === 'move'
+            )
+
+            expect(additionalInfoRow.actions.items).to.eql([
+              {
+                attributes: {
+                  'data-update-link': 'move',
+                },
+                category: 'move',
+                href: 'http://www.example.com/move-details',
+                html: 'moves::update_link.link_text',
+              },
+            ])
+          })
+        })
+      })
+
+      context('when not a prison recall move', function () {
+        beforeEach(function () {
+          transformedResponse = moveToAdditionalInfoListComponent(
+            {
+              ...mockMove,
+              time_due: null,
+            },
+            updateUrls
+          )
+        })
+
+        describe('response', function () {
+          it('should not include the update link for additional info', function () {
+            const additionalInfoRow = transformedResponse.rows.find(
+              row => row.updateJourneyKey === 'move'
+            )
+
+            expect(additionalInfoRow).to.be.undefined
+          })
         })
       })
     })
@@ -156,7 +267,7 @@ describe('Presenters', function () {
 
       validMoveTypes.forEach(moveType => {
         it('should return correct keys', function () {
-          transformedResponse = moveToSummaryListComponent({
+          transformedResponse = moveToAdditionalInfoListComponent({
             ...mockMove,
             move_type: moveType,
           })
@@ -173,7 +284,7 @@ describe('Presenters', function () {
 
       invalidMoveTypes.forEach(moveType => {
         it('should return undefined', function () {
-          transformedResponse = moveToSummaryListComponent({
+          transformedResponse = moveToAdditionalInfoListComponent({
             ...mockMove,
             move_type: moveType,
           })
