@@ -6,12 +6,12 @@ import { PersonEscortRecord } from './person-escort-record'
 
 export interface Event {
   id: string
-  event_type: string
-  classification: string
+  event_type?: string
+  classification?: string
   occurred_at: string
-  recorded_at: string
-  notes: string | null
-  created_by: string | null
+  recorded_at?: string
+  notes?: string | null
+  created_by?: string | null
   eventable?: Journey | Move | PersonEscortRecord | Person
   supplier?: string
   details: EventDetails
